not sure if this is everything needed for auth

+++ routes/api.php
@@ -18,35 +18,42 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 // USERS
-Route::get('users', 'UserController@index');
-Route::get('users/{id}', 'UserController@show');
-Route::post('users', 'UserController@store');
-Route::put('users/{id}', 'UserController@update');
-Route::delete('users/{id}', 'UserController@delete');
+Route::middleware('auth:api')->get('users', 'UserController@index');
+Route::middleware('auth:api')->get('users/{id}', 'UserController@show');
+Route::middleware('auth:api')->post('users', 'UserController@store');
+Route::middleware('auth:api')->put('users/{id}', 'UserController@update');
+Route::middleware('auth:api')->delete('users/{id}', 'UserController@delete');
 // TASK LISTS
-Route::get('task_lists', 'TaskListController@index');
-Route::get('task_lists/{id}', 'TaskListController@show');
-Route::post('task_lists', 'TaskListController@store');
-Route::put('task_lists/{id}', 'TaskListController@update');
-Route::delete('task_lists/{id}', 'TaskListController@delete');
+Route::middleware('auth:api')->get('task_lists', 'TaskListController@index');
+Route::middleware('auth:api')->get('task_lists/{id}', 'TaskListController@show');
+Route::middleware('auth:api')->post('task_lists', 'TaskListController@store');
+Route::middleware('auth:api')->put('task_lists/{id}', 'TaskListController@update');
+Route::middleware('auth:api')->delete('task_lists/{id}', 'TaskListController@delete');
 // TASKS
-Route::get('tasks', 'TaskController@index');
-Route::get('tasks/{id}', 'TaskController@show');
-Route::post('tasks', 'TaskController@store');
-Route::put('tasks/{id}', 'TaskController@update');
-Route::delete('tasks/{id}', 'TaskController@delete');
+Route::middleware('auth:api')->get('tasks', 'TaskController@index');
+Route::middleware('auth:api')->get('tasks/{id}', 'TaskController@show');
+Route::middleware('auth:api')->post('tasks', 'TaskController@store');
+Route::middleware('auth:api')->put('tasks/{id}', 'TaskController@update');
+Route::middleware('auth:api')->delete('tasks/{id}', 'TaskController@delete');
 // TASK STATUS
-Route::get('task_status', 'TaskStatusController@index');
-Route::get('task_status/{id}', 'TaskStatusController@show');
-Route::post('task_status', 'TaskStatusController@store');
-Route::put('task_status/{id}', 'TaskStatusController@update');
+Route::middleware('auth:api')->get('task_status', 'TaskStatusController@index');
+Route::middleware('auth:api')->get('task_status/{id}', 'TaskStatusController@show');
+Route::middleware('auth:api')->post('task_status', 'TaskStatusController@store');
+Route::middleware('auth:api')->put('task_status/{id}', 'TaskStatusController@update');
 // TASK PRIORITIES
-Route::get('task_priorities', 'TaskPriorityController@index');
-Route::get('task_priorities/{id}', 'TaskPriorityController@show');
-Route::post('task_priorities', 'TaskPriorityController@store');
-Route::put('task_priorities/{id}', 'TaskPriorityController@update');
+Route::middleware('auth:api')->get('task_priorities', 'TaskPriorityController@index');
+Route::middleware('auth:api')->get('task_priorities/{id}', 'TaskPriorityController@show');
+Route::middleware('auth:api')->post('task_priorities', 'TaskPriorityController@store');
+Route::middleware('auth:api')->put('task_priorities/{id}', 'TaskPriorityController@update');
 // PRIORITY CODES
-Route::get('priority_codes', 'PriorityCodeController@index');
-Route::get('priority_codes/{id}', 'PriorityCodeController@show');
-Route::post('priority_codes', 'PriorityCodeController@store');
-Route::put('priority_codes/{id}', 'PriorityCodeController@update');
+Route::middleware('auth:api')->get('priority_codes', 'PriorityCodeController@index');
+Route::middleware('auth:api')->get('priority_codes/{id}', 'PriorityCodeController@show');
+Route::middleware('auth:api')->post('priority_codes', 'PriorityCodeController@store');
+Route::middleware('auth:api')->put('priority_codes/{id}', 'PriorityCodeController@update');
+// AUTHORIZATION
+// Authentication routes...
+Route::post('auth/login', 'AuthController@postLogin');
+Route::middleware('auth:api')->get('auth/logout', 'AuthController@postLogout');
+
+// Registration routes...
+Route::post('auth/register', 'AuthController@create');
